Move kvec.rs to vec.rs and use `Memory` where possible
