+++ src/common/src/vec.rs
@@ -1,13 +1,11 @@
 use core::clone::Clone;
 use core::iter::Iterator;
-use core::mem::size_of;
 use core::ops::Drop;
 use core::option::Option;
 use core::ptr;
-use core::slice;
-use core::slice::SliceExt;
+use core::slice::{self, SliceExt};
 
-use syscall::call::*;
+use common::memory::*;
 
 #[macro_export]
 macro_rules! vec {
@@ -38,109 +36,94 @@ impl <'a, T> Iterator for VecIterator<'a, T> {
     }
 }
 
-/// An owned iterator over a vec
-pub struct OwnedVecIterator<T>
-    where T: Clone
-{
-    vec: Vec<T>,
-    offset: usize,
-}
-
-impl<'a, T> Iterator for OwnedVecIterator<T>
-            where T: Clone {
-    type Item = T;
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.vec.get(self.offset) {
-            Option::Some(item) => {
-                self.offset += 1;
-                Option::Some(item.clone())
-            }
-            Option::None => {
-                Option::None
-            }
-        }
-    }
-}
 /// A owned, heap allocated list of elements
 pub struct Vec<T> {
-    pub data: *mut T,
+    pub mem: Memory<T>, // TODO: Option<Memory>
     pub length: usize,
 }
 
 impl <T> Vec<T> {
     /// Create a empty vector
     pub fn new() -> Self {
-        Vec::<T> {
-            data: 0 as *mut T,
+        Vec {
+            mem: Memory { ptr: 0 as *mut T /* TODO: Option::None */ },
             length: 0,
         }
     }
 
     /// Convert to pointer
     pub unsafe fn as_ptr(&self) -> *const T {
-        self.data
+        self.mem.ptr
     }
 
     /// Convert from a raw (unsafe) buffer
     pub unsafe fn from_raw_buf(ptr: *const T, len: usize) -> Self {
-        let data = sys_alloc(size_of::<T>() * len) as *mut T;
+        match Memory::new(len) {
+            Option::Some(mem) => {
+                ptr::copy(ptr, mem.ptr, len);
 
-        ptr::copy(ptr, data, len);
-
-        Vec::<T> {
-            data: data,
-            length: len,
+                return Vec {
+                    mem: mem,
+                    length: len,
+                };
+            }
+            Option::None => {
+                return Self::new();
+            }
         }
     }
 
     pub fn from_slice(slice: &[T]) -> Self {
-        let data;
-        unsafe {
-            data = sys_alloc(size_of::<T>() * slice.len()) as *mut T;
+        match Memory::new(slice.len()) {
+            Option::Some(mem) => {
+                unsafe { ptr::copy(slice.as_ptr(), mem.ptr, slice.len()) };
 
-            ptr::copy(slice.as_ptr(), data, slice.len());
-        }
-
-        Vec::<T> {
-            data: data,
-            length: slice.len(),
+                return Vec {
+                    mem: mem,
+                    length: slice.len(),
+                };
+            }
+            Option::None => {
+                return Vec::new();
+            }
         }
     }
+
 
     /// Get the nth element. Returns None if out of bounds.
     pub fn get(&self, i: usize) -> Option<&mut T> {
         if i >= self.length {
             Option::None
         } else {
-            unsafe { Option::Some(&mut *self.data.offset(i as isize)) }
+            unsafe { Option::Some(&mut *self.mem.ptr.offset(i as isize)) }
         }
     }
 
     /// Set the nth element
     pub fn set(&self, i: usize, value: T) {
         if i <= self.length {
-            unsafe {
-                ptr::write(self.data.offset(i as isize), value);
-            }
+            unsafe { ptr::write(self.mem.ptr.offset(i as isize), value) };
         }
     }
 
     /// Insert element at a given position
     pub fn insert(&mut self, i: usize, value: T) {
         if i <= self.length {
-            self.length += 1;
-            unsafe {
-                self.data = sys_realloc(self.data as usize, self.length * size_of::<T>()) as *mut T;
+            let new_length = self.length + 1;
+            if self.mem.renew(new_length) {
+                self.length = new_length;
 
                 //Move all things ahead of insert forward one
                 let mut j = self.length - 1;
                 while j > i {
-                    ptr::write(self.data.offset(j as isize),
-                               ptr::read(self.data.offset(j as isize - 1)));
+                    unsafe {
+                        ptr::write(self.mem.ptr.offset(j as isize),
+                                   ptr::read(self.mem.ptr.offset(j as isize - 1)));
+                    }
                     j -= 1;
                 }
 
-                ptr::write(self.data.offset(i as isize), value);
+                unsafe { ptr::write(self.mem.ptr.offset(i as isize), value) };
             }
         }
     }
@@ -149,21 +132,22 @@ impl <T> Vec<T> {
     pub fn remove(&mut self, i: usize) -> Option<T> {
         if i < self.length {
             self.length -= 1;
-            unsafe {
-                let item = ptr::read(self.data.offset(i as isize));
 
-                //Move all things ahead of remove back one
-                let mut j = i;
-                while j < self.length {
-                    ptr::write(self.data.offset(j as isize),
-                               ptr::read(self.data.offset(j as isize + 1)));
-                    j += 1;
+            let item = unsafe { ptr::read(self.mem.ptr.offset(i as isize)) };
+
+            //Move all things ahead of remove back one
+            let mut j = i;
+            while j < self.length {
+                unsafe {
+                    ptr::write(self.mem.ptr.offset(j as isize),
+                               ptr::read(self.mem.ptr.offset(j as isize + 1)));
                 }
-
-                self.data = sys_realloc(self.data as usize, self.length * size_of::<T>()) as *mut T;
-
-                Option::Some(item)
+                j += 1;
             }
+
+            self.mem.renew(self.length);
+
+            Option::Some(item)
         } else {
             Option::None
         }
@@ -171,10 +155,11 @@ impl <T> Vec<T> {
 
     /// Push an element to a vector
     pub fn push(&mut self, value: T) {
-        self.length += 1;
-        unsafe {
-            self.data = sys_realloc(self.data as usize, self.length * size_of::<T>()) as *mut T;
-            ptr::write(self.data.offset(self.length as isize - 1), value);
+        let new_length = self.length + 1;
+        if self.mem.renew(new_length) {
+            self.length = new_length;
+
+            unsafe { ptr::write(self.mem.ptr.offset(self.length as isize - 1), value) };
         }
     }
 
@@ -182,15 +167,15 @@ impl <T> Vec<T> {
     pub fn pop(&mut self) -> Option<T> {
         if self.length > 0 {
             self.length -= 1;
-            unsafe {
-                let item = ptr::read(self.data.offset(self.length as isize));
-                self.data = sys_realloc(self.data as usize, self.length * size_of::<T>()) as *mut T;
 
-                Option::Some(item)
-            }
-        } else {
-            Option::None
+            let item = unsafe { ptr::read(self.mem.ptr.offset(self.length as isize)) };
+
+            self.mem.renew(self.length);
+
+            return Option::Some(item);
         }
+
+        Option::None
     }
 
     /// Get the length of the vector
@@ -201,16 +186,6 @@ impl <T> Vec<T> {
     /// Create an iterator
     pub fn iter(&self) -> VecIterator<T> {
         VecIterator {
-            vec: self,
-            offset: 0,
-        }
-    }
-
-    /// Convert into an iterator
-    pub fn into_iter(self) -> OwnedVecIterator<T>
-        where T: Clone
-    {
-        OwnedVecIterator {
             vec: self,
             offset: 0,
         }
@@ -233,24 +208,29 @@ impl <T> Vec<T> {
             return Vec::new();
         }
 
-        unsafe {
-            let data = sys_alloc(length * size_of::<T>()) as *mut T;
+        match Memory::new(length) {
+            Option::Some(mem) => {
+                for k in i..j {
+                    unsafe {
+                        ptr::write(mem.ptr.offset((k - i) as isize),
+                                   ptr::read(self.mem.ptr.offset(k as isize)))
+                    };
+                }
 
-            for k in i..j {
-                ptr::write(data.offset((k - i) as isize),
-                           ptr::read(self.data.offset(k as isize)));
+                return Vec {
+                    mem: mem,
+                    length: length,
+                };
             }
-
-            Vec {
-                data: data,
-                length: length,
+            Option::None => {
+                return Self::new();
             }
         }
     }
 
     pub fn as_slice(&self) -> &[T] {
-        if self.data as usize > 0 && self.length > 0 {
-            unsafe { slice::from_raw_parts(self.data, self.length) }
+        if self.length > 0 {
+            unsafe { slice::from_raw_parts(self.mem.ptr, self.length) }
         } else {
             &[]
         }
@@ -261,12 +241,12 @@ impl<T> Vec<T> where T: Clone {
     /// Append a vector to another vector
     pub fn push_all(&mut self, vec: &Self) {
         let mut i = self.length as isize;
-        self.length += vec.len();
-        unsafe {
-            self.data = sys_realloc(self.data as usize, self.length * size_of::<T>()) as *mut T;
+        let new_length = self.length + vec.len();
+        if self.mem.renew(new_length) {
+            self.length = new_length;
 
             for value in vec.iter() {
-                ptr::write(self.data.offset(i), value.clone());
+                unsafe { ptr::write(self.mem.ptr.offset(i), value.clone()) };
                 i += 1;
             }
         }
@@ -275,7 +255,7 @@ impl<T> Vec<T> where T: Clone {
 
 impl<T> Clone for Vec<T> where T: Clone {
     fn clone(&self) -> Self {
-        let mut ret = Vec::new();
+        let mut ret = Self::new();
         ret.push_all(self);
         ret
     }
@@ -285,12 +265,8 @@ impl<T> Drop for Vec<T> {
     fn drop(&mut self) {
         unsafe {
             for i in 0..self.len() {
-                ptr::read(self.data.offset(i as isize));
+                ptr::read(self.mem.ptr.offset(i as isize));
             }
-
-            sys_unalloc(self.data as usize);
-            self.data = 0 as *mut T;
-            self.length = 0;
         }
     }
 }
