+++ src/kernel.rs
@@ -75,7 +75,6 @@ mod common {
     pub mod debug;
     pub mod elf;
     pub mod event;
-    pub mod kvec;
     pub mod queue;
     pub mod memory;
     pub mod mutex;
